ER-14 Move ZodError.ts to lib folder, update import paths
